chore: 配置优化与平台管理功能更新
- 优化 Claude settings 配置
- 更新 Cursor rules 配置文件
- 添加平台管理 i18n 翻译 (zh-CN/en-US)
- 新增工厂管理和状态机设计器页面
- 更新意图配置清理功能

+++ backend-java/src/main/java/com/cretas/aims/dto/intent/CleanupRequest.java
@@ -4,6 +4,8 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 
+import java.math.BigDecimal;
+
 /**
  * 清理低效关键词的请求DTO
  */
@@ -15,10 +17,10 @@ public class CleanupRequest {
     /**
      * Wilson Score 阈值，低于此值的关键词将被清理
      */
-    private Double threshold;
+    private BigDecimal threshold;
 
     /**
      * 最小负面反馈次数，达到此次数的关键词将被清理
      */
-    private Integer minNegative;
+    private int minNegative;
 }
